Suitable method of re-using child views
Could also add more declarations in the HTML, such as:

<div data-name=“ctlrPropertyName” data-view=“ClassOfView”></div>

This would be sufficient to eliminate the @views property of the controller... but that does make child views a little less consistent with other Spine idioms, such as @events and @elements.

Could also look at creating a View class that dealt specifically with adding behavior to DOM structures and rendering templates, but that would require a bastardization of Spine.Controller.